Change v0 into v1
