--- fabric-environment-attributes-v0/src/testmod/resources/fabric.mod.json
@@ -1,19 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-environment-attributes-v0-testmod",
-  "name": "Fabric Environment Attributes (v0) Test Mod",
-  "version": "1.0.0",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "depends": {
-    "fabric-environment-attributes-v0": "*"
-  },
-  "entrypoints": {
-    "main": [
-      "net.fabricmc.fabric.test.environment.attribute.FabricEnvironmentAttributesTest"
-    ],
-    "fabric-client-gametest": [
-      "net.fabricmc.fabric.test.environment.attribute.client.FabricEnvironmentAttributesClientTest"
-    ]
-  }
-}
