+++ fabric-environment-attributes-v1/build.gradle
