+++ fabric-environment-attributes-v1/src/main/resources/assets/fabric-environment-attributes-v1/icon.png
