+++ fabric-environment-attributes-v1/src/main/resources/fabric-environment-attributes-v1.mixins.json
