+++ fabric-environment-attributes-v1/src/main/resources/fabric.mod.json
@@ -1,0 +1,31 @@
+{
+  "schemaVersion": 1,
+  "id": "fabric-environment-attributes-v1",
+  "name": "Fabric Environment Attributes API (v1)",
+  "version": "${version}",
+  "license": "Apache-2.0",
+  "icon": "assets/fabric-environment-attributes-v1/icon.png",
+  "contact" : {
+    "homepage": "https://fabricmc.net",
+    "irc": "irc://irc.esper.net:6667/fabric",
+    "issues": "https://github.com/FabricMC/fabric/issues",
+    "sources": "https://github.com/FabricMC/fabric"
+  },
+  "authors": [
+    "FabricMC"
+  ],
+  "entrypoints": {
+  },
+  "depends": {
+    "fabricloader": ">=0.18.4",
+    "minecraft": ">=1.16-rc.3",
+    "fabric-api-base": "*"
+  },
+  "description": "Fabric Environment Attributes API.",
+  "mixins": [
+    "fabric-environment-attributes-v1.mixins.json"
+  ],
+  "custom": {
+    "fabric-api:module-lifecycle": "stable"
+  }
+}
