+++ gradle.properties
@@ -25,7 +25,7 @@ fabric-data-generation-api-v1-version=24.0.9
 fabric-debug-api-v1-version=1.0.0
 fabric-dimensions-v1-version=5.1.1
 fabric-entity-events-v1-version=5.0.0
-fabric-environment-attributes-v0-version=1.0.0
+fabric-environment-attributes-v1-version=1.0.0
 fabric-events-interaction-v0-version=5.1.4
 fabric-game-rule-api-v1-version=4.0.3
 fabric-gametest-api-v1-version=4.0.6
