+++ settings.gradle
@@ -40,7 +40,7 @@ include 'fabric-data-generation-api-v1'
 include 'fabric-debug-api-v1'
 include 'fabric-dimensions-v1'
 include 'fabric-entity-events-v1'
-include 'fabric-environment-attributes-v0'
+include 'fabric-environment-attributes-v1'
 include 'fabric-events-interaction-v0'
 include 'fabric-game-rule-api-v1'
 include 'fabric-gametest-api-v1'
